Add jquery-highlight dependency to vendor imports
It looks like this has defined the highlight code to be added to the vendor tree, but it isn't actually included as an import when `vendor.js` is created. This should add the required code to the `vendor.js` output. This appears to fix the issues I described in #1. Sorry for hijacking that issue.